Build: enable -DOGR_ENABLE_DRIVER_GML_PLUGIN=ON, provided that -DOGR_ENABLE_DRIVER_NAS=OFF; make CSW and WFS drivers plugin capable too
