+++ frmts/ogcapi/gdalogcapidataset.cpp
@@ -35,9 +35,7 @@
 #include "ogrsf_frmts.h"
 #include "ogr_spatialref.h"
 
-#ifdef OGR_ENABLE_DRIVER_GML
 #include "parsexsd.h"
-#endif
 
 #include <algorithm>
 #include <memory>
@@ -1649,7 +1647,6 @@ GDALColorInterp OGCAPIMapWrapperBand::GetColorInterpretation()
 /*                           ParseXMLSchema()                           */
 /************************************************************************/
 
-#ifdef OGR_ENABLE_DRIVER_GML
 static bool
 ParseXMLSchema(const std::string &osURL,
                std::vector<std::unique_ptr<OGRFieldDefn>> &apoFields,
@@ -1694,7 +1691,6 @@ ParseXMLSchema(const std::string &osURL,
 
     return false;
 }
-#endif
 
 /************************************************************************/
 /*                         InitWithTilesAPI()                           */
@@ -1966,14 +1962,12 @@ bool OGCAPIDataset::InitWithTilesAPI(GDALOpenInfo *poOpenInfo,
             }
         }
 
-#ifdef OGR_ENABLE_DRIVER_GML
         std::vector<std::unique_ptr<OGRFieldDefn>> apoFields;
         bool bGotSchema = false;
         if (!osXMLSchemaURL.empty())
         {
             bGotSchema = ParseXMLSchema(osXMLSchemaURL, apoFields, eGeomType);
         }
-#endif
 
         for (const auto &tileMatrix : tms->tileMatrixList())
         {
@@ -2021,10 +2015,8 @@ bool OGCAPIDataset::InitWithTilesAPI(GDALOpenInfo *poOpenInfo,
                     tileMatrix, eGeomType));
             poLayer->SetMinMaxXY(minCol, minRow, maxCol, maxRow);
             poLayer->SetExtent(dfXMin, dfYMin, dfXMax, dfYMax);
-#ifdef OGR_ENABLE_DRIVER_GML
             if (bGotSchema)
                 poLayer->SetFields(apoFields);
-#endif
             m_apoLayers.emplace_back(std::move(poLayer));
         }
