+++ ogr/ogrsf_frmts/csw/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_gdal_driver(
   TARGET ogr_CSW
   SOURCES ogrcswdataset.cpp
-  BUILTIN) # BUILTIN because of GML and WFS dependency
+  PLUGIN_CAPABLE
+  NO_DEPS)
 gdal_standard_includes(ogr_CSW)
-target_include_directories(ogr_CSW PRIVATE $<TARGET_PROPERTY:ogr_GML,SOURCE_DIR> $<TARGET_PROPERTY:ogr_WFS,SOURCE_DIR>)
+target_include_directories(ogr_CSW PRIVATE $<TARGET_PROPERTY:gmlutils,SOURCE_DIR>)
