+++ ogr/ogrsf_frmts/gmlutils/gmlpropertydefn.cpp
@@ -28,7 +28,7 @@
  ****************************************************************************/
 
 #include "cpl_port.h"
-#include "gmlreader.h"
+#include "gmlfeature.h"
 
 #include <cstring>
