+++ ogr/ogrsf_frmts/gmlutils/gmlutils.h
@@ -50,24 +50,25 @@ typedef enum
     SRSNAME_OGC_URL
 } OGRGMLSRSNameFormat;
 
-const char *
+const char CPL_DLL *
 GML_ExtractSrsNameFromGeometry(const CPLXMLNode *const *papsGeometry,
                                std::string &osWork, bool bConsiderEPSGAsURN);
 
-bool GML_IsSRSLatLongOrder(const char *pszSRSName);
-bool GML_IsLegitSRSName(const char *pszSRSName);
+bool CPL_DLL GML_IsSRSLatLongOrder(const char *pszSRSName);
+bool CPL_DLL GML_IsLegitSRSName(const char *pszSRSName);
 
-void *GML_BuildOGRGeometryFromList_CreateCache();
-void GML_BuildOGRGeometryFromList_DestroyCache(void *hCacheSRS);
+void CPL_DLL *GML_BuildOGRGeometryFromList_CreateCache();
+void CPL_DLL GML_BuildOGRGeometryFromList_DestroyCache(void *hCacheSRS);
 
-OGRGeometry *GML_BuildOGRGeometryFromList(
+OGRGeometry CPL_DLL *GML_BuildOGRGeometryFromList(
     const CPLXMLNode *const *papsGeometry, bool bTryToMakeMultipolygons,
     bool bInvertAxisOrderIfLatLong, const char *pszDefaultSRSName,
     bool bConsiderEPSGAsURN, GMLSwapCoordinatesEnum eSwapCoordinates,
     int nPseudoBoolGetSecondaryGeometryOption, void *hCacheSRS,
     bool bFaceHoleNegative = false);
 
-char *GML_GetSRSName(const OGRSpatialReference *poSRS,
-                     OGRGMLSRSNameFormat eSRSNameFormat, bool *pbCoordSwap);
+char CPL_DLL *GML_GetSRSName(const OGRSpatialReference *poSRS,
+                             OGRGMLSRSNameFormat eSRSNameFormat,
+                             bool *pbCoordSwap);
 
 #endif /* _CPL_GMLREADERP_H_INCLUDED */
