+++ ogr/ogrsf_frmts/nas/CMakeLists.txt
@@ -9,7 +9,7 @@ add_gdal_driver(
           ogrnaslayer.cpp
   BUILTIN) # BUILTIN because of GML dependency
 gdal_standard_includes(ogr_NAS)
-target_include_directories(ogr_NAS PRIVATE $<TARGET_PROPERTY:ogr_GML,SOURCE_DIR>)
+target_include_directories(ogr_NAS PRIVATE $<TARGET_PROPERTY:ogr_GML,SOURCE_DIR> $<TARGET_PROPERTY:gmlutils,SOURCE_DIR>)
 if (GDAL_USE_EXPAT)
   target_compile_definitions(ogr_NAS PRIVATE -DHAVE_EXPAT)
   gdal_target_link_libraries(ogr_NAS PRIVATE ${EXPAT_TARGET})
