+++ ogr/ogrsf_frmts/wfs/ogrwfsjoinlayer.cpp
@@ -27,6 +27,7 @@
  ****************************************************************************/
 
 #include "ogr_wfs.h"
+#include "ogrwfsfilter.h"
 #include "cpl_md5.h"
 
 /************************************************************************/
